+++ cashu-mint-rest-it/src/test/java/xyz/tcheeric/cashu/mint/rest/interop/NutshellInteropIT.java
@@ -34,6 +34,7 @@ import xyz.tcheeric.cashu.mint.proto.service.MintLoadService;
 import xyz.tcheeric.cashu.mint.proto.service.MintProtocolService;
 import xyz.tcheeric.cashu.mint.proto.service.MintVaultService;
 import xyz.tcheeric.cashu.mint.proto.service.ProofVaultService;
+import xyz.tcheeric.cashu.mint.rest.support.ProofVaultStubs;
 import xyz.tcheeric.cashu.mint.proto.service.impl.MintProtocolServiceFactory;
 import xyz.tcheeric.cashu.mint.rest.spec001.AbstractMintDurableIT;
 import xyz.tcheeric.cashu.mint.rest.spec002.support.MeltProofFixture;
@@ -161,6 +162,7 @@ class NutshellInteropIT extends AbstractMintDurableIT {
 
     @BeforeEach
     void publishKeysetAndStubVault() throws Exception {
+        ProofVaultStubs.keyProofsByIssuanceKey(proofVaultService);
         Mint mint = MeltProofFixture.mintWithKeys();
         when(mintLoadService.load(any(UUID.class), Mockito.anyBoolean())).thenReturn(mint);
         when(mintLoadService.load(Mockito.anyBoolean())).thenReturn(List.of(mint));